backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
torch==2.2.2
torchvision==0.17.2
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
transformers==4.41.1
tokenizers>=0.15.0
safetensors>=0.4.2
